Allow setting the password via the web UI
So I can use this on mobile.

+++ index.js
@@ -3,8 +3,7 @@ const itemUl = document.querySelector('#itemUl');
 
 const password = localStorage.getItem('password');
 if (!password) {
-  textInput.disabled = true;
-  textInput.placeholder = 'Set `password` in `localStorage`';
+  textInput.placeholder = 'Submit to set password';
 }
 
 textInput.addEventListener('keydown', async function (event) {
@@ -17,6 +16,12 @@ textInput.addEventListener('keydown', async function (event) {
     return;
   }
 
+  if (!password) {
+    localStorage.setItem('password', text);
+    location.reload();
+    return;
+  }
+
   await fetch(`/${password}`, { method: 'POST', body: text });
   textInput.value = '';
   location.reload();
@@ -24,43 +29,45 @@ textInput.addEventListener('keydown', async function (event) {
 
 textInput.focus();
 
-const response = await fetch(`/${password}`);
-const items = await response.json();
+if (password) {
+  const response = await fetch(`/${password}`);
+  const items = await response.json();
 
-for (const item of items) {
-  const li = document.createElement('li');
+  for (const item of items) {
+    const li = document.createElement('li');
 
-  const deleteButton = document.createElement('button');
-  deleteButton.textContent = '✕';
-  li.append(deleteButton);
+    const deleteButton = document.createElement('button');
+    deleteButton.textContent = '✕';
+    li.append(deleteButton);
 
-  deleteButton.addEventListener('click', async function () {
-    if (!confirm(`Delete item "${item.name}"?`)) {
-      return;
-    }
+    deleteButton.addEventListener('click', async function () {
+      if (!confirm(`Delete item "${item.name}"?`)) {
+        return;
+      }
 
-    await fetch(`/${password}?name=${item.name}`, { method: 'DELETE' });
-    li.remove();
-  });
-
-  li.append(` ${item.name}: `);
-
-  const textSpan = document.createElement('span');
-  textSpan.textContent = item.text;
-  li.append(textSpan);
-
-  textSpan.addEventListener('click', async function () {
-    const newText = prompt('Text:', item.text);
-    if (newText === null || newText === item.text) {
-      return;
-    }
-
-    await fetch(`/${password}?name=${item.name}`, {
-      method: 'PUT',
-      body: newText,
+      await fetch(`/${password}?name=${item.name}`, { method: 'DELETE' });
+      li.remove();
     });
-    textSpan.textContent = newText;
-  });
 
-  itemUl.append(li);
+    li.append(` ${item.name}: `);
+
+    const textSpan = document.createElement('span');
+    textSpan.textContent = item.text;
+    li.append(textSpan);
+
+    textSpan.addEventListener('click', async function () {
+      const newText = prompt('Text:', item.text);
+      if (newText === null || newText === item.text) {
+        return;
+      }
+
+      await fetch(`/${password}?name=${item.name}`, {
+        method: 'PUT',
+        body: newText,
+      });
+      textSpan.textContent = newText;
+    });
+
+    itemUl.append(li);
+  }
 }
